Fix too wide math accents on scripted symbols
With OpenType math fonts, accents over scripted glyphs are too wide as
if the accent were to be placed over the scripts as well. This because
XeTeX is choosing the accent based on the width of the box after adding
the scripts to it unlike how it does with legacy fonts.

Now we use the width of the box before adding the scripts to it.