Fix vertical shift of OT math scripts
Since we the parameter fetching macros to always use the current font,
they are now affected by the fact that in make_scripts current font
changes to the script or scriptscript font which gives use scaled down
values, so now I reset cur_f to make sure it points to the correct font.

Rather and ugly hack, but should do the trick for now.