Don’t support using “()” for quoting filenames
This was added back in 2004 for use with ConTeXt as an alternative to
using double and quotes, for quoting native fonts (which is scanned by
by the same routine as file name), but it made it not possible to use
file names with parenthesis.

http://tug.org/pipermail/xetex/2004-October/001215.html

It seems nobody is actually using that feature, it is not documented and
no other engine supports it, so I’m dropping it until somebody
complains.