Drop TeX--XeT code
There is a serious issue with TeX--XeT; right to left “hlist” is reversed
in “hlist_out” to make the text right to left, but this also reverses
“special” nodes and as result all nodes that donate begin/end (e.g.
color or PDF annotation) gets reversed and thus applied in the wrong
order.

To fix this, instead of reversion the node order we should only change
their x coordinates so that they are laid out in the reverse order
visually while still in the correct order in the DVI file (this what
Omega did).

The next commit will use the new approach.