Apply original Knuth and MacKay TeX-XeT
Unlike eTeX’s TeX--XeT, the original code does not do node reversal, it
just inserts “begin_reflect” and “end_reflect” DVI opcode and delegates
actual text reversal to the driver. This shouldn’t be an issue for us
since we require a special DVI driver already.

The driver (xdvipdfmx in our case) does not actually reverse the order
of the text, it instead changes its coordinates, as a result “special”
nodes are kept in their logical order which allows color and annotations
to work correctly.

TODO:
- handling of right to left lines interrupted by displays
- \predisplaydirection?