Return nothing if no match was found in getFileNameFromCTFont()
The Helvetica CY font is broken, it has PS name like HelveticaCYPlain
(which is wrong), but CT is trying to be smart and "corrects" it into
HelveticaCY-Plain, but this breaks our code. Furthermore, it has a
seemingly broken 'post' table for it, so we cannot correctly use it any
way.