Fix x_offset in vertical mode
To move right in real vertical mode is actually to move upwards in
xetex, so it seems that we have to reverse the plus sign of x_offset.
In fact I know a font that has vertical gpos feature using x_offset,
and this font works correctly when the following patch has been
applied.  In the same logic, it seems that the sign of x_advance
should be reversed as well.  But I am not sure as I do not have any
font to test this.