Fix Y positioning with AAT fonts
We were ignoring the Y position since d32d45c as it were breaking fake
slanting, but this breaks any AAT font with Y positions (like Farisi and
other new Arabic fonts on Mac). It turns out the real bug was that we
were applying the slanting in the Y instead of X direction!