Avoid prematurely merging following text with the no-break text of a \discretionary.
We don't want to combine text after \discretionary{A}{B}{C} into the same node
as the (trailing node of the) no-break parameter {C}, because this would result
in following text being lost in the case where the discretionary break is taken.
When this happens, its no-break text (recorded as a count of nodes) is supposed
to be skipped over, but if we've combined following characters with the trailing
node, they will also be (incorrectly) skipped.

Fixes #137.