Fix the code to skip discretionaries when looking for a |native_word| node immediately preceding the current |tail|.
This has been broken for years, see #149.

This reverts part of 6a6805694f5509fe1c08ef2669ec3f514bc1649d (the addition
of the |disc| flag to |native_word| nodes), which is no longer needed now
that the code here works as intended.

Because this did not work properly, we had buggy behavior with discretionary
nodes, resulting in issue #137, where nodes could be discarded after a
discretionary break. That was fixed by the rather cumbersome addition of a
special flag to |native_word| nodes that belong to a discretionary, but should
not have been a problem in the first place if this code had been correct.